Se creo autenticación por medio de cookies para el usuario pueda ver solo si existe en la base de datos, a su vez inicialmente solo puede ver la información de su apartamento

+++ ClienteWeb/Controllers/ApartamentosController.cs
@@ -16,12 +16,37 @@ namespace ClienteWeb.Controllers
 
         //Listar Apartamentos
         [ResponseCache(NoStore = true, Location = ResponseCacheLocation.None)]
+        public async Task<IActionResult> IndexAll()
+        {
+            if (!Request.Cookies.TryGetValue("UsuarioResponsable", out string usuario))
+            {
+                return RedirectToAction("Login", "Auth");
+            }
+            var apartamento = await _apiService.GetApartamentosAsync();
+            if (apartamento == null)
+            {
+                return RedirectToAction("Logout", "Auth");
+            }
+            return View(apartamento);
+        }
+        //Listar Apto
         public async Task<IActionResult> Index()
         {
-            var apartamentos = await _apiService.GetApartamentosAsync();
-            return View(apartamentos);
-        }
+            var usuarioResponsable = Request.Cookies["UsuarioResponsable"];
+            if (string.IsNullOrEmpty(usuarioResponsable))
+            {
+                return RedirectToAction("Login", "Auth");
+            }
 
+            var apartamento = await _apiService.GetApartamentoPorUsuario(usuarioResponsable);
+            if (apartamento == null)
+            {
+                TempData["Error"] = "No tienes apartamento asignado.";
+                return RedirectToAction("Login", "Auth");
+            }
+
+            return View(apartamento);
+        }
         //Creación
         public IActionResult Create() => View();
 
@@ -29,17 +54,22 @@ namespace ClienteWeb.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(ApartamentoViewModel model)
         {
-
+            if (!Request.Cookies.TryGetValue("UsuarioResponsable", out string usuario))
+            {
+                return RedirectToAction("Login", "Auth");
+            }
             if (ModelState.IsValid &&  await _apiService.CreateApartamentoAsync(model))
                 return RedirectToAction("Index");
-                
-
             return View(model);
         }
 
         //Edit
         public async Task<IActionResult> Edit (int id)
         {
+            if (!Request.Cookies.TryGetValue("UsuarioResponsable", out string usuario))
+            {
+                return RedirectToAction("Login", "Auth");
+            }
             var apartamento = await _apiService.GetApartamentoByIdAsync(id);
             return apartamento == null ? NotFound() : View(apartamento);
         }
@@ -59,6 +89,10 @@ namespace ClienteWeb.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!Request.Cookies.TryGetValue("UsuarioResponsable", out string usuario))
+            {
+                return RedirectToAction("Login", "Auth");
+            }
             var response = await _apiService.DeleteApartamentoAsync(id);
 
             if (!response)
